+++ timetypes/doc.go
@@ -1,5 +1,5 @@
-// Package rfc3339type implements a terraform-plugin-framework attr.Type and
+// package timetypes implements a terraform-plugin-framework attr.Type and
 // attr.Value for [RFC 3339] timestamp strings.
 //
 // [RFC 3339]: https://tools.ietf.org/html/rfc3339
-package rfc3339type
+package timetypes