+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# 0.2.0 (October 3, 2022)
+
+BREAKING CHANGES:
+
+* The `rfc3339type` package was moved to `timetypes`, the `Type` type to `RFC3339Type`, and the `Value` type to `RFC3339`.
+
+NOTES:
+
+* This Go module was moved to `github.com/bflad/terraform-plugin-framework-type-time`.
+
 # 0.1.0 (September 22, 2022)
 
 FEATURES: